ajoit du cout total par mois+ qq ameliorations

+++ proto/xmart-city-business-dto/src/main/java/edu/ezip/ing1/pds/business/dto/TotalCout.java
@@ -33,7 +33,7 @@ public class TotalCout {
         this.dateAchat = dateAchat;
     }
 
-    public int getTotalCout() {
+    public double getTotalCout() {
         return totalCout;
     }
