+++ proto/xmart-city-business-dto/src/main/java/edu/ezip/ing1/pds/business/dto/TotalMaintenance.java
@@ -33,7 +33,7 @@ public class TotalMaintenance {
         this.dateMaintenance = dateMaintenance;
     }
 
-    public int getTotalMaintenance() {
+    public double getTotalMaintenance() {
         return totalMaintenance;
     }
